+++ jopa-impl/src/main/java/cz/cvut/kbss/jopa/query/criteria/RootImpl.java
@@ -6,14 +6,24 @@ import cz.cvut.kbss.jopa.model.query.criteria.Root;
 import cz.cvut.kbss.jopa.query.criteria.expressions.AbstractPathExpression;
 import cz.cvut.kbss.jopa.sessions.CriteriaBuilder;
 
-public class RootImpl<X> extends PathImpl<X> implements Root<X> {
+public class RootImpl<X> extends AbstractPathExpression<X> implements Root<X> {
 
     public RootImpl(Metamodel metamodel, AbstractPathExpression<X> expression, Class<X> type, CriteriaBuilder cb) {
-        super(metamodel, expression, type, cb);
+        super(type, expression, metamodel, cb);
     }
 
     @Override
     public EntityType<X> getModel() {
         return metamodel.entity(type);
+    }
+
+    @Override
+    public void setExpressionToQuery(StringBuilder query, CriteriaParameterFiller parameterFiller) {
+        if (this.pathSource != null){
+            this.pathSource.setExpressionToQuery(query, parameterFiller);
+            query.append("." + type.getSimpleName().toLowerCase());
+        } else {
+            query.append(type.getSimpleName().toLowerCase());
+        }
     }
 }
